+++ turbopuffer-java-core/src/test/kotlin/com/turbopuffer/services/async/NamespaceServiceAsyncTest.kt
@@ -223,6 +223,7 @@ internal class NamespaceServiceAsyncTest {
             namespaceServiceAsync.write(
                 NamespaceWriteParams.builder()
                     .namespace("namespace")
+                    .branchFromNamespace("branch_from_namespace")
                     .copyFromNamespace("string")
                     .addDelete("182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e")
                     .addDelete("182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e")